rename package external to serviceentry
