+++ pilot/pkg/serviceregistry/serviceentry/conversion.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package external
+package serviceentry
 
 import (
 	"net"
